# Conflicts:
#	user-service/src/main/java/rs/edu/raf/banka1/bootstrap/BootstrapData.java
#	user-service/src/main/java/rs/edu/raf/banka1/configuration/SpringSecurityConfig.java

+++ user-service/src/main/java/rs/edu/raf/banka1/utils/Constants.java
@@ -3,5 +3,5 @@ package rs.edu.raf.banka1.utils;
 public class Constants {
     public static final String ALL_POSITIONS = "All";
     public static final Integer BEARER_PREFIX_SIZE = 7;
-    public static final Integer JWT_EXPIRATION_LENGTH = 1000 * 60 * 60 * 10;
+    public static final Long JWT_EXPIRATION_LENGTH = 1000L * 60 * 60 * 10;
 }
